Moving iree/base/wait_handle.h to iree/base/internal/. We'll expose the concept through dedicated APIs where it makes sense (like importing/exporting platform handles/etc) vs. allowing users to poke right into these things.
