+++ iree/base/BUILD
@@ -257,33 +257,3 @@ cc_library(
         "@com_google_absl//absl/base:core_headers",
     ],
 )
-
-cc_library(
-    name = "wait_handle",
-    srcs = [
-        "wait_handle.c",
-        "wait_handle_epoll.c",
-        "wait_handle_impl.h",
-        "wait_handle_kqueue.c",
-        "wait_handle_poll.c",
-        "wait_handle_posix.c",
-        "wait_handle_posix.h",
-        "wait_handle_win32.c",
-    ],
-    hdrs = ["wait_handle.h"],
-    deps = [
-        ":api",
-        ":core_headers",
-        ":tracing",
-    ],
-)
-
-cc_test(
-    name = "wait_handle_test",
-    srcs = ["wait_handle_test.cc"],
-    deps = [
-        ":wait_handle",
-        "//iree/testing:gtest",
-        "//iree/testing:gtest_main",
-    ],
-)
