+++ iree/base/internal/wait_handle_kqueue.c
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 // NOTE: must be first to ensure that we can define settings for all includes.
-#include "iree/base/wait_handle_impl.h"
+#include "iree/base/internal/wait_handle_impl.h"
 
 #if IREE_WAIT_API == IREE_WAIT_API_KQUEUE
 
+#include "iree/base/internal/wait_handle_posix.h"
 #include "iree/base/tracing.h"
-#include "iree/base/wait_handle_posix.h"
 
 //===----------------------------------------------------------------------===//
 // iree_wait_set_t
