+++ iree/base/internal/wait_handle_posix.c
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iree/base/wait_handle_posix.h"
+#include "iree/base/internal/wait_handle_posix.h"
 
 #include "iree/base/tracing.h"
