+++ iree/base/internal/wait_handle_win32.c
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // NOTE: must be first to ensure that we can define settings for all includes.
-#include "iree/base/wait_handle_impl.h"
+#include "iree/base/internal/wait_handle_impl.h"
 
 #if defined(IREE_PLATFORM_WINDOWS)
