+++ iree/hal/local/BUILD
@@ -50,7 +50,8 @@ cc_library(
         "//iree/base:core_headers",
         "//iree/base:synchronization",
         "//iree/base:tracing",
-        "//iree/base:wait_handle",
+        "//iree/base/internal",
+        "//iree/base/internal:wait_handle",
     ],
 )
 
@@ -115,8 +116,8 @@ cc_library(
         "//iree/base:core_headers",
         "//iree/base:synchronization",
         "//iree/base:tracing",
-        "//iree/base:wait_handle",
         "//iree/base/internal",
+        "//iree/base/internal:wait_handle",
         "//iree/hal:api",
         "//iree/task",
     ],
