+++ iree/hal/local/CMakeLists.txt
@@ -39,9 +39,10 @@ iree_cc_library(
   DEPS
     iree::base::api
     iree::base::core_headers
+    iree::base::internal
+    iree::base::internal::wait_handle
     iree::base::synchronization
     iree::base::tracing
-    iree::base::wait_handle
   PUBLIC
 )
 
@@ -106,9 +107,9 @@ iree_cc_library(
     iree::base::api
     iree::base::core_headers
     iree::base::internal
+    iree::base::internal::wait_handle
     iree::base::synchronization
     iree::base::tracing
-    iree::base::wait_handle
     iree::hal::api
     iree::task
   PUBLIC
