+++ iree/task/BUILD
@@ -54,10 +54,10 @@ cc_library(
         "//iree/base:synchronization",
         "//iree/base:threading",
         "//iree/base:tracing",
-        "//iree/base:wait_handle",
         "//iree/base/internal",
         "//iree/base/internal:atomic_slist",
         "//iree/base/internal:prng",
+        "//iree/base/internal:wait_handle",
         "@cpuinfo",
     ],
 )
